refactor: Change stof to stoi

+++ src/main_ut.cpp
@@ -36,7 +36,7 @@ public:
         getline(in_stream, x);
         istringstream iss(x);
         vector<string> x2(istream_iterator<string>{iss}, istream_iterator<string>());
-        in_data.walk_ways_number = stof(x2[0]);
+        in_data.walk_ways_number = stoi(x2[0]);
         in_data.distance_to_walk= stof(x2[1]);
 
         while(getline(in_stream, x)&&(x != "")){
